2h: add grid to page:

+++ SimpleExcelViewer.Domain/Repositories/IManagerRepository.cs
@@ -7,13 +7,13 @@ using System.Threading.Tasks;
 
 namespace SimpleExcelViewer.Domain.Repositories
 {
-	public interface IManagerRepository: IRepository<Manager, long>
+	public interface IManagerRepository: IRepository<Manager, int>
 	{
 		/// <summary>
 		/// Returns all managers from one <see cref="Office"/> by specified <see cref="Office.Id"/>
 		/// </summary>
 		/// <param name="officeId"><see cref="Office.Id"/></param>
 		/// <returns>Query with collection of <see cref="Manager"/></returns>
-		IQueryable<Manager> GetManagersFromOffice(long officeId);
+		IQueryable<Manager> GetManagersFromOffice(int officeId);
 	}
 }
