+++ SimpleExcelViewer.Domain/Repositories/OfficeRepository.cs
@@ -8,13 +8,13 @@ namespace SimpleExcelViewer.Domain.Repositories
 	/// <summary>
 	/// Implementation of <see cref="IRepository{T, TId}"/> for <see cref="Office"/>
 	/// </summary>
-	public class OfficeRepository : BaseRepository<Office, long>, IOfficeRepository
+	public class OfficeRepository : BaseRepository<Office, int>, IOfficeRepository
 	{
 		public OfficeRepository(IDataSet<Office> dataSet) : base(dataSet)
 		{
 		}
 
-		public IQueryable<Office> GetAllOfficesForManager(long managerId)
+		public IQueryable<Office> GetAllOfficesForManager(int managerId)
 		{
 			throw new NotImplementedException();
 		}
